Cleanup/warning fixes

+++ src/imgui_backend/ImguiShaderCompiler.cpp
@@ -39,7 +39,7 @@ void NOINLINE ReadCompiledShader(GLSLCoutput *compileData) {
 
     Logger::log("Shader Count: %d\n", compileData->numSections);
 
-    for (int i = 0; i < compileData->numSections; ++i) {
+    for (uint32_t i = 0; i < compileData->numSections; ++i) {
         if (compileData->headers[i].genericHeader.common.type != GLSLCsectionTypeEnum::GLSLC_SECTION_TYPE_GPU_CODE)
             continue;
         auto compInfo = &compileData->headers[i].gpuCodeHeader;
@@ -64,7 +64,7 @@ CompiledData NOINLINE CreateShaderBinary(GLSLCoutput *compileData, const char *s
 
     size_t binarySize = 0;
 
-    for (int i = 0; i < compileData->numSections; ++i) {
+    for (uint32_t i = 0; i < compileData->numSections; ++i) {
         if (compileData->headers[i].genericHeader.common.type == GLSLCsectionTypeEnum::GLSLC_SECTION_TYPE_GPU_CODE) {
             binarySize = ALIGN_UP(binarySize + compileData->headers[i].genericHeader.common.size, 0x100);
         }
@@ -81,7 +81,7 @@ CompiledData NOINLINE CreateShaderBinary(GLSLCoutput *compileData, const char *s
     u32 headerInfo[4] = {};
 
     // place control sections first (as we dont need to align it)
-    for (int i = 0; i < compileData->numSections; ++i) {
+    for (uint32_t i = 0; i < compileData->numSections; ++i) {
         if (compileData->headers[i].genericHeader.common.type != GLSLCsectionTypeEnum::GLSLC_SECTION_TYPE_GPU_CODE)
             continue;
 
@@ -94,7 +94,7 @@ CompiledData NOINLINE CreateShaderBinary(GLSLCoutput *compileData, const char *s
     }
 
     // place data sections next
-    for (int i = 0; i < compileData->numSections; ++i) {
+    for (uint32_t i = 0; i < compileData->numSections; ++i) {
         if (compileData->headers[i].genericHeader.common.type != GLSLCsectionTypeEnum::GLSLC_SECTION_TYPE_GPU_CODE)
             continue;
         auto compInfo = &compileData->headers[i].gpuCodeHeader;
@@ -167,8 +167,8 @@ bool ImguiShaderCompiler::CheckIsValidVersion(nvn::Device *device) {
         device->GetInteger(nvn::DeviceInfo::GLSLC_MAX_SUPPORTED_GPU_CODE_MINOR_VERSION, &maxMinorVersion);
         Logger::log("NVN Api Max Minor Version: %d\n", maxMinorVersion);
 
-        if ((versionInfo.apiMajor >= minMajorVersion && versionInfo.apiMajor <= maxMajorVersion) &&
-            (versionInfo.apiMinor >= minMinorVersion && versionInfo.apiMinor <= maxMinorVersion)) {
+        if ((versionInfo.apiMajor >= static_cast<uint32_t>(minMajorVersion) && versionInfo.apiMajor <= static_cast<uint32_t>(maxMajorVersion)) &&
+            (versionInfo.apiMinor >= static_cast<uint32_t>(minMinorVersion) && versionInfo.apiMinor <= static_cast<uint32_t>(maxMinorVersion))) {
             Logger::log("NVN Api supports GLSLC version!\n");
             return true;
         } else if (minMajorVersion == 1 && maxMajorVersion == 1) {
